Entity design and relationship management

+++ src/main/java/org/launchcode/vendormangedinventory/models/daos/PurchaseDao.java
@@ -7,5 +7,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 @Transactional
-public interface PurchaseDao extends CrudRepository<Purchase, Integer> {
+public interface PurchaseDao extends CrudRepository<Purchase, Long> {
 }
